Refactor output_collect into galaxy-tool-util.
Needed for dynamic discovery of outputs on remote servers.
