+++ lib/galaxy/tool_util/output_collect.py
